컴포넌트 리펙토링
공통 코드 분리
