+++ src/components/ISwiperProps.tsx
@@ -1,0 +1,10 @@
+export type ISwiperProps = { 
+  loop?: boolean
+  spaceBetween?: number 
+  scrollbar?: { 
+    draggable: boolean 
+    el: null 
+  }
+  slidesPerView?: "auto" | number | undefined
+  onSlideChange?: (swiper: any) => void
+}
