+++ src/pages/home/Profile.tsx
@@ -3,33 +3,17 @@ import React, { useState } from "react"
 import NavigationBar from "../../components/NavigationBar"
 import { jamesLee } from "../../assets/images/brand"
 import { IconHeartBig, IconHeartBigRed } from "../../assets/images"
-import {Paper, IconButton, Typography, Box, FormControl, Tabs, Tab } from '@mui/material'
-import { styled } from "@mui/material/styles"
+import { IconButton, Typography, Box, Tabs, Tab } from '@mui/material'
 import DetailViewPanel from "./DetailViewPanel"
 import BrandShopPanel from "./BrandShopPanel"
+import {RectCheckboxNonChecked} from "../../styles/checkbox.styles"
+import {profileTagData} from "../../data"
 
 interface TabPanelProps {
   children?: React.ReactNode
   index: number
   value: number
 }
-
-const RectCheckbox = styled(FormControl)(({ }) => ({
-  '& input': {
-    display: 'none',
-  },
-  '& label': {
-    padding: '6px 16px',
-    border: '1px solid #DADADA',
-    fontSize: '0.875rem',
-    color: '#999999',
-  },
-  '& input:checked + label': {
-    border: '1px solid #242223',
-    backgroundColor: '#242223',
-    color: '#fff',
-  },
-}))
 
 function profileProps(index: number) {
   return {
@@ -40,7 +24,6 @@ function profileProps(index: number) {
 
 function LikeTabPanel(props: TabPanelProps) {
   const { children, value, index, ...other } = props
-
   return (
     <div
       role="tabpanel"
@@ -58,48 +41,15 @@ function LikeTabPanel(props: TabPanelProps) {
   )
 }
 
-const itemList = [
-  {
-    id: '1',
-    name: '금속디자인',
-  },
-  {
-    id: '2',
-    name: '미드센추리',
-  },
-  {
-    id: '3',
-    name: '스테인리스 스틸',
-  },
-  {
-    id: '4',
-    name: '반짝이는 가구',
-  },
-  {
-    id: '5',
-    name: '금속 가구',
-  },
-  {
-    id: '6',
-    name: '북유럽',
-  },
-]
-
 const Profile = () => {
-
   const [tabValue, setTabValue] = useState(0)
   const [isLike, setIsLike] = useState(false)
-
   const handleLikeTab = (event:React.SyntheticEvent, newValue: number) => {
     setTabValue(newValue)
   }
-
   const handleClickLike = (event: React.MouseEvent) => {
     setIsLike(!isLike)
   }
-
-  
-  
   return (
     <>
       <img src={jamesLee} style={{width: "100vw", maxWidth: "420px"}}/>
@@ -118,14 +68,14 @@ const Profile = () => {
         </Box>
       </Box>
       <Box sx={{my:4}}>
-      {itemList.map((el) => (
-          <RectCheckbox key={el.id} sx={{ ml: 1, mb: 1 }}>
+      {profileTagData.map((el) => (
+          <RectCheckboxNonChecked key={el.id} sx={{ ml: 1, mb: 1 }}>
             <input
               type="checkbox"
               id={`check_${el.id}`}
             />
             <label htmlFor={`check_${el.id}`}>{el.name}</label>
-          </RectCheckbox>
+          </RectCheckboxNonChecked>
         ))}
       </Box>
       <Box sx={{
