+++ src/pages/like/Like/BrandList.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import { Typography, Avatar, List, ListItemAvatar, ListItemButton, ListItemText } from "@mui/material"
 import { styled } from "@mui/material/styles"
 import { HalfArrow } from "../../../assets/images"
-import { aerobiey, alfdn, clods, jameslee } from "../../../assets/images/logo"
+import { brandListData } from "../../../data"
 import { useNavigate } from 'react-router-dom'
 
 const AvatarList = styled(List)(() => ({
@@ -17,32 +17,7 @@ const AvatarList = styled(List)(() => ({
     fontSize: '.875rem',
   },
 }))
-
-const itemList = [
-  {
-    id: 'james_lee',
-    name: 'James Lee',
-    logo: jameslee
-  },
-  {
-    id: 'alfdn',
-    name: 'ALFDN',
-    logo: alfdn
-  },
-  {
-    id: 'rerobiey',
-    name: 'Rerobiey',
-    logo: aerobiey
-  },
-  {
-    id: 'clods',
-    name: 'CLODS',
-    logo: clods
-  }
-]
-
 const BrandList = () => {
-
   const navigate = useNavigate();
   const goToBrand = () => {
     navigate("/style/brand")
@@ -50,9 +25,9 @@ const BrandList = () => {
 
   return (
     <>
-      <Typography sx={{ py: 3, px: 2, fontSize: '0.875rem' }}>{itemList.length+"개의 브랜드"}</Typography>
+      <Typography sx={{ py: 3, px: 2, fontSize: '0.875rem' }}>{brandListData.length+"개의 브랜드"}</Typography>
       <AvatarList sx={{ py: 0 }}>
-      {itemList.map((el, index) => (
+      {brandListData.map((el, index) => (
           <ListItemButton
           key={el.id}
           onClick={() => goToBrand()}
