+++ src/pages/like/Like/ProductList.tsx
@@ -1,127 +1,50 @@
-import React, { useState } from "react"
-import { Box, FormControl, FormControlLabel, Checkbox, Button } from "@mui/material"
-import { styled } from "@mui/material/styles"
+import React, { useState, useEffect } from "react"
+import { Box, FormControlLabel, Button } from "@mui/material"
 import { IconShift } from "../../../assets/images"
 import List from "./List"
-import { chair03, sofa01, sofa02, sofa03 } from "../../../assets/images/product"
-
-const RectCheckbox = styled(FormControl)(({ }) => ({
-  '& input': {
-    display: 'none',
-  },
-  '& label': {
-    padding: '3px 12px',
-    border: '1px solid #DADADA',
-    fontSize: '0.875rem',
-  },
-  '& input:checked + label': {
-    border: '1px solid #242223',
-    backgroundColor: '#242223',
-    color: '#fff',
-  },
-}))
-
-const itemList = [
-  {
-    id: 'sofa',
-    name: '소파',
-  },
-  {
-    id: 'bed',
-    name: '침대',
-  },
-  {
-    id: 'desk',
-    name: '책상',
-  },
-  {
-    id: 'table',
-    name: '테이블',
-  },
-  {
-    id: 'chair',
-    name: '의자',
-  },
-  {
-    id: 'lights',
-    name: '조명',
-  },
-  {
-    id: 'kids',
-    name: '키즈',
-  },
-  {
-    id: 'closet',
-    name: '수납장',
-  },
-  {
-    id: 'home_deco',
-    name: '홈데코',
-  }
-]
-
-export const testData = [
-  {
-    id: 0,
-    imgsrc: chair03,
-    like: true,
-    likeCount: 556,
-    shopName: 'Henredo',
-    modelName: '[23s] HRD-0152',
-    productName: '릴렉스 체어 (3colors)',
-    price: 4834000,
-    colors: ['#A55555', '#192552', '#494A4C']
-  },
-  {
-    id: 1,
-    imgsrc: sofa02,
-    like: false,
-    likeCount: 556,
-    shopName: 'Henredo',
-    modelName: '[22s] HRD-210',
-    productName: '1인 모던 쇼파 ',
-    price: 6790000,
-    colors: ['#864000']
-  },
-  {
-    id: 2,
-    imgsrc: sofa03,
-    like: false,
-    likeCount: 556,
-    shopName: 'Henredo',
-    modelName: '[23s] HRD-251',
-    productName: '기획전 시티 쇼파',
-    price: 2230000,
-    colors: ['#CBC6C3']
-  },
-  {
-    id: 3,
-    imgsrc: sofa01,
-    like: false,
-    likeCount: 556,
-    shopName: 'Henredo',
-    modelName: '[23s] HRD-123',
-    productName: '올드 가죽쇼파',
-    price: 1230000,
-    colors: ['#A55555']
-  },
-]
+import { CustomCheckbox } from "../../../styles/checkbox.styles"
+import {likeProductData, styleTypes} from "../../../data/index"
+import {RectCheckbox} from "../../../styles/checkbox.styles"
 
 const ProductList = () => {
   const [filterItem, setFilterItem] = useState<string[]>([])
+  const [isAllChecked, setIsAllChecked] = useState(true);
+  const [filteredStyles, setFilteredStyles] = useState(likeProductData);
 
-  const handleCheckedFilterItem = (e: React.ChangeEvent<HTMLInputElement>, id: string) => {
-    if (id === 'all') {
+  const itemList = styleTypes.map(style => ({
+    id: (styleTypes.indexOf(style) + 1).toString(),
+    name: style.name,
+  }));
+
+  useEffect(() => {
+    setFilterItem(itemList.map(el => el.name));
+  }, []);
+
+  useEffect(() => {
+    const updatedFilteredStyles = likeProductData.filter((style) => filterItem.includes(style.name));
+    setFilteredStyles(updatedFilteredStyles);
+  }, [filterItem]);
+
+  useEffect(() => {
+    if (filterItem.length === styleTypes.length) {
+      setIsAllChecked(true);
+    } else {
+      setIsAllChecked(false);
+    }
+  }, [filterItem]);
+
+  const handleCheckedFilterItem = (e: React.ChangeEvent<HTMLInputElement>, name: string) => {
+    if (name === 'all') {
       if (e.currentTarget.checked) {
-        setFilterItem(itemList.map(el => el.id))
+        setFilterItem(itemList.map(el => el.name))
       } else {
         setFilterItem([])
       }
     } else {
       if (e.currentTarget.checked) {
-        setFilterItem([...filterItem, id])
+        setFilterItem([...filterItem, name])
       } else {
-        setFilterItem(filterItem.filter(el => el !== id))
+        setFilterItem(filterItem.filter(el => el !== name))
       }
     }
   }
@@ -142,6 +65,7 @@ const ProductList = () => {
             type="checkbox"
             id={`check_all`}
             onChange={e => handleCheckedFilterItem(e, 'all')}
+            checked={isAllChecked}
           />
           <label htmlFor={`check_all`}>전체</label>
         </RectCheckbox>
@@ -149,11 +73,11 @@ const ProductList = () => {
           <RectCheckbox key={el.id} sx={{ ml: 1 }}>
             <input
               type="checkbox"
-              id={`check_${el.id}`}
-              onChange={e => handleCheckedFilterItem(e, el.id)}
-              checked={filterItem.includes(el.id)}
+              id={el.name}
+              onChange={e => handleCheckedFilterItem(e, el.name)}
+              checked={filterItem.includes(el.name)}
             />
-            <label htmlFor={`check_${el.id}`}>{el.name}</label>
+            <label htmlFor={`check_${el.name}`}>{el.name}</label>
           </RectCheckbox>
         ))}
       </Box>
@@ -169,14 +93,8 @@ const ProductList = () => {
         }}
       >
         <FormControlLabel
-          control={<Checkbox color="default" sx={{ p: '0px 4px 0px 8px' }} />}
+          control={<CustomCheckbox type="checkbox" style={{marginLeft:"12px", marginRight: "8px"}}/>}
           label="품절 제외"
-          sx={{
-            '& > .MuiTypography-root': {
-              color: '#333',
-              fontSize: '.875rem'
-            }
-          }}
         />
         <Button
           sx={{
@@ -193,7 +111,7 @@ const ProductList = () => {
         </Button>
       </Box>
       <Box>
-        <List data={testData} />
+        <List data={filteredStyles} />
       </Box>
     </>
   )
