+++ src/pages/mypage/Coupon.tsx
@@ -2,63 +2,19 @@ import React, { useState, useEffect  } from "react"
 import { Box, Tabs, Tab, Typography, Button } from "@mui/material"
 import NavigationBar from "../../components/NavigationBar"
 import CenterTitle from "../../components/title/CenterTitle"
+import { hasCouponData, getCouponData} from "../../data"
+import { TabPanel, TabProps} from "../../components/Tabs"
 import { HalfArrowDownGray } from "../../assets/images"
-import { onerain_coupon03 } from "../../assets/images/product"
 import CouponCard from "./CouponCard"
 
-interface TabPanelProps {
-  children?: React.ReactNode
-  index: number
-  value: number
-}
-
-const data = [
-  {
-    imgsrc: onerain_coupon03,
-    shopName: 'Aerobiey',
-    couponName: '에어로우뷔 브랜드 10% 쿠폰',
-    startDate: '20230925',
-    endDate: '20231120',
-  },
-  {
-    imgsrc: onerain_coupon03,
-    shopName: 'James Lee',
-    couponName: '제임스 리  브랜드 10% 쿠폰',
-    startDate: '20230920',
-    endDate: '20231115',
-  },
-  {
-    imgsrc: onerain_coupon03,
-    shopName: 'One rain',
-    couponName: '원레인  브랜드 10% 쿠폰',
-    startDate: '20230922',
-    endDate: '20231128',
-  },
-]
-
-const data2 = [
-  {
-    imgsrc: onerain_coupon03,
-    shopName: 'James Lee',
-    couponName: '제임스 리  브랜드 10% 쿠폰',
-    startDate: '20230920',
-    endDate: '20231118',
-  },
-  {
-    imgsrc: onerain_coupon03,
-    shopName: 'Aerobiey',
-    couponName: '에어로우뷔 브랜드 10% 쿠폰',
-    startDate: '20230925',
-    endDate: '20231118',
-  },
-  {
-    imgsrc: onerain_coupon03,
-    shopName: 'One rain',
-    couponName: '원레인  브랜드 10% 쿠폰',
-    startDate: '20230822',
-    endDate: '20231128',
-  },
-]
+type CouponData = {
+  imgsrc: string;
+  shopName: string;
+  couponName: string;
+  startDate: string;
+  endDate: string;
+  dDay: number;
+};
 
 const formatDate = (dateString: string) => {
   const year = dateString.slice(0, 4);
@@ -78,7 +34,8 @@ const calculateDDay = (endDate: string): number => {
   return daysDiff;
 };
 
-const formattedData = data.map((item) => ({
+//보유 쿠폰 D-day 추가한 데이터 초기화
+const formattedHasCouponData = hasCouponData.map((item) => ({
   imgsrc: item.imgsrc,
   shopName: item.shopName,
   couponName: item.couponName,
@@ -87,7 +44,8 @@ const formattedData = data.map((item) => ({
   dDay: calculateDDay(item.endDate),
 }));
 
-const formattedData2 = data2.map((item) => ({
+//쿠폰 빋기 D-day 추가한 데이터 초기화
+const formattedGetCouponData = getCouponData.map((item) => ({
   imgsrc: item.imgsrc,
   shopName: item.shopName,
   couponName: item.couponName,
@@ -95,64 +53,25 @@ const formattedData2 = data2.map((item) => ({
   endDate: formatDate(item.endDate),
   dDay: calculateDDay(item.endDate),
 }));
-
-type CouponData = {
-  imgsrc: string;
-  shopName: string;
-  couponName: string;
-  startDate: string;
-  endDate: string;
-  dDay: number;
-};
-
-function couponProps(index: number) {
-  return {
-    id: `coupon-tab-${index}`,
-    'aria-controls': `coupon-tabpanel-${index}`,
-  }
-}
-
-function CouponTabPanel(props: TabPanelProps) {
-  const { children, value, index, ...other } = props
-
-  return (
-    <div
-      role="tabpanel"
-      hidden={value !== index}
-      id={`coupon-tabpanel-${index}`}
-      aria-labelledby={`coupon-tab-${index}`}
-      {...other}
-    >
-      {value === index && (
-        <Box>
-          {children}
-        </Box>
-      )}
-    </div>
-  )
-}
-
-const title = ['쿠폰'];
 
 const Coupon: React.FC = () => {
   const [tabValue, setTabValue] = useState(0);
-  const [dataPanel1, setDataPanel1] = useState<CouponData[]>(formattedData);
-  const [sortedDataPanel1, setSortedDataPanel1] = useState<CouponData[]>([]);
-  const [isSortedPanel1, setIsSortedPanel1] = useState(false);
-  const [dataPanel2, setDataPanel2] = useState<CouponData[]>(formattedData2);
-  const [sortedDataPanel2, setSortedDataPanel2] = useState<CouponData[]>([]);
-  const [isSortedPanel2, setIsSortedPanel2] = useState(false);
+  const [hasCouponDataPanel, setHasCouponDataPanel] = useState<CouponData[]>(formattedHasCouponData);
+  const [sortedHasCouponDataPanel, setSortedHasCouponDataPanel] = useState<CouponData[]>([]);
+  const [isSortedHasCouponPanel, setIsSortedHasCouponPanel] = useState(false);
+  const [getCouponDataPanel, setGetCouponDataPanel] = useState<CouponData[]>(formattedGetCouponData);
+  const [sortedGetCouponDataPanel, setSortedGetCouponDataPanel] = useState<CouponData[]>([]);
+  const [isSortedGetCouponPanel, setIsSortedGetCouponPanel] = useState(false);
   const [couponCount, setCouponCount] = useState(0);
-
 
   const handleCouponTab = (event:React.SyntheticEvent, newValue: number) => {
     setTabValue(newValue)
   }
 
   useEffect(() => {
-    setDataPanel1(formattedData);
-    setDataPanel2(formattedData2);
-    setCouponCount(formattedData2.length);
+    setHasCouponDataPanel(formattedHasCouponData);
+    setGetCouponDataPanel(formattedGetCouponData);
+    setCouponCount(formattedGetCouponData.length);
   }, []);
 
   const sortData = (unsortedData: CouponData[]) => {
@@ -161,22 +80,23 @@ const Coupon: React.FC = () => {
   };
 
   const handleCouponClick = (index: number) => {
-    if (isSortedPanel2) {
-      const deletedValue = sortedDataPanel2[index];
-      const updatedSortedData = sortedDataPanel2.filter((item) => item !== deletedValue);
-      const updatedUnsortedData = dataPanel2.filter((item) => item !== deletedValue);
-      setSortedDataPanel2(updatedSortedData);
-      setDataPanel2(updatedUnsortedData);
+    if (isSortedGetCouponPanel) {
+      const deletedValue = sortedGetCouponDataPanel[index];
+      const updatedSortedData = sortedGetCouponDataPanel.filter((item) => item !== deletedValue);
+      const updatedUnsortedData = getCouponDataPanel.filter((item) => item !== deletedValue);
+      setSortedGetCouponDataPanel(updatedSortedData);
+      setGetCouponDataPanel(updatedUnsortedData);
       setCouponCount(updatedUnsortedData.length);
     } else {
-      const deletedValue = dataPanel2[index];
-      const updatedData = dataPanel2.filter((item) => item !== deletedValue);
-      setSortedDataPanel2(updatedData);
-      setDataPanel2(updatedData);
+      const deletedValue = getCouponDataPanel[index];
+      const updatedData = getCouponDataPanel.filter((item) => item !== deletedValue);
+      setSortedGetCouponDataPanel(updatedData);
+      setGetCouponDataPanel(updatedData);
       setCouponCount(updatedData.length);
     }
   };
 
+  const title = ['쿠폰'];
   return (
     <>
       <CenterTitle title={title}/>
@@ -193,24 +113,24 @@ const Coupon: React.FC = () => {
             onChange={handleCouponTab} 
             aria-label="main category"
           >
-            <Tab label="보유 쿠폰" {...couponProps(0)} />
-            <Tab label="쿠폰 받기" {...couponProps(1)} />
+            <Tab label="보유 쿠폰" {...TabProps(0)} />
+            <Tab label="쿠폰 받기" {...TabProps(1)} />
           </Tabs>
         </Box>
 
-        <CouponTabPanel value={tabValue} index={0}>
+        <TabPanel value={tabValue} index={0}>
           <Box sx={{mx:"20px", mt:6}}>
             <Box sx={{display:"flex", alignItems:"center", justifyContent:"space-between", mb:"21px"}}>
-              <Typography sx={{fontSize: "12px", lineHeight: "12px", letterSpacing: "0.4px", color:"#333333", fontWeight:"500"}}>{"쿠폰 "+formattedData.length+"개"}</Typography>
+              <Typography sx={{fontSize: "12px", lineHeight: "12px", letterSpacing: "0.4px", color:"#333333", fontWeight:"500"}}>{"쿠폰 "+formattedHasCouponData.length+"개"}</Typography>
               <Box sx={{display:"flex"}}>
                 <Button sx={{color:"#333", display:"flex", borderRadius:0}} onClick={() => {
-                  const sortedDataCopy = sortData(dataPanel1);
-                  if (!isSortedPanel1) {
-                    setSortedDataPanel1(sortedDataCopy);
-                    setIsSortedPanel1(true);
+                  const sortedDataCopy = sortData(hasCouponDataPanel);
+                  if (!isSortedHasCouponPanel) {
+                    setSortedHasCouponDataPanel(sortedDataCopy);
+                    setIsSortedHasCouponPanel(true);
                   } else {
-                    setDataPanel1(sortedDataCopy.reverse());
-                    setIsSortedPanel1(false);
+                    setHasCouponDataPanel(sortedDataCopy.reverse());
+                    setIsSortedHasCouponPanel(false);
                   }
                 }}>
                   <Typography sx={{fontSize: "12px", lineHeight: "16px", color:"#999999", fontWeight:"400"}}>최신순</Typography>
@@ -220,27 +140,27 @@ const Coupon: React.FC = () => {
             </Box>
             
             {/* 쿠폰 */}
-            {isSortedPanel1 ? (
-              <CouponCard data={sortedDataPanel1}/>
+            {isSortedHasCouponPanel ? (
+              <CouponCard data={sortedHasCouponDataPanel}/>
             ) : (
-              <CouponCard data={dataPanel1}/>
+              <CouponCard data={hasCouponDataPanel}/>
             )}
           </Box>
-        </CouponTabPanel>
+        </TabPanel>
 
-        <CouponTabPanel value={tabValue} index={1}>
+        <TabPanel value={tabValue} index={1}>
           <Box sx={{mx:"20px", mt:6}}>
             <Box sx={{display:"flex", alignItems:"center", justifyContent:"space-between", mb:"21px"}}>
               <Typography sx={{fontSize: "12px", lineHeight: "12px", letterSpacing: "0.4px", color:"#333333", fontWeight:"500"}}>{"쿠폰 "+couponCount+"개"}</Typography>
               <Box sx={{display:"flex"}}>
                 <Button sx={{color:"#333", display:"flex", borderRadius:0}} onClick={() => {
-                  const sortedDataCopy = sortData(dataPanel2);
-                  if (!isSortedPanel2) {
-                    setSortedDataPanel2(sortedDataCopy);
-                    setIsSortedPanel2(true);
+                  const sortedDataCopy = sortData(getCouponDataPanel);
+                  if (!isSortedGetCouponPanel) {
+                    setSortedGetCouponDataPanel(sortedDataCopy);
+                    setIsSortedGetCouponPanel(true);
                   } else {
-                    setDataPanel2(sortedDataCopy.reverse());
-                    setIsSortedPanel2(false);
+                    setGetCouponDataPanel(sortedDataCopy.reverse());
+                    setIsSortedGetCouponPanel(false);
                   }
                 }}>
                 <Typography sx={{fontSize: "12px", lineHeight: "16px", color:"#999999", fontWeight:"400"}}>최신순</Typography>
@@ -250,13 +170,13 @@ const Coupon: React.FC = () => {
             </Box>
             
             {/* 쿠폰 */}
-            {isSortedPanel2 ? (
-              <CouponCard data={sortedDataPanel2} onCouponClick={(index) => handleCouponClick(index)} />
+            {isSortedGetCouponPanel ? (
+              <CouponCard data={sortedGetCouponDataPanel} onCouponClick={(index) => handleCouponClick(index)} />
             ) : (
-              <CouponCard data={dataPanel2}onCouponClick={(index) => handleCouponClick(index)} />
+              <CouponCard data={getCouponDataPanel}onCouponClick={(index) => handleCouponClick(index)} />
             )}
           </Box>
-        </CouponTabPanel>
+        </TabPanel>
       </Box>
       <NavigationBar />
     </>
