+++ src/pages/mypage/DeliveryDetail.tsx
@@ -1,100 +1,15 @@
 import React, { useState } from "react"
-import { Box, Card, CardHeader, CardMedia, CardActions, Typography, Button } from "@mui/material"
+import { Box, CardHeader, CardMedia, CardActions, Typography, Button } from "@mui/material"
 import NavigationBar from "../../components/NavigationBar"
 import CenterTitle from "../../components/title/CenterTitle"
-import { deliveryProduct } from "../../assets/images/product"
-import { styled } from "@mui/material/styles" 
+import { myDeliveryInfo } from "../../data"
 import { OutlineButton } from "../../styles/buttons.styles"
+import { DeliveryInfoOrderCard } from "../../styles/card.style"
 import { IconCircleGray, IconDeliveryCompleted, IconTelGray } from "../../assets/images"
-
-const OrderCard = styled(Card)(() => ({
-  marginBottom: '16px',
-  borderRadius: 0,
-  border: '0.5px solid #8F8F8F', 
-  boxShadow: 'none', 
-  '& .MuiCardHeader-root': {
-    padding: '8px 16px',
-    borderBottom: '0.5px solid #8F8F8F',
-    '& .MuiCardHeader-title': {
-      color: '#666',
-      fontSize: '.875rem',
-    },
-  },
-  '& .MuiCardMedia-root': {
-    marginRight: '16px',
-    flex: '0 0 100px',
-  },
-  '& .card_content': {
-    padding: '16px',
-  },
-  '& .card_content .state': {
-    fontSize: '.75rem', 
-    color: '#333', 
-    fontWeight: 'bold',
-  },
-  '& .card_content .product': {
-    display: 'flex',
-    marginTop: '16px',
-  },
-  '& .MuiCardActions-root': {
-    padding: '16px',
-    '& > button': {
-      flex: '1 1 auto',
-      borderRadius: 0,
-      borderColor: '#8F8F8F',
-      color: '#666',
-    },
-  },
-}))
 
 const centerTitle = ['배송 조회']
 
-const MyDeliveryInfo = [
-  {
-    date: '2023.02.23',
-    state: '배송준비중',
-    img: deliveryProduct,
-    brand: 'HC ART',
-    optionCode: '[21Series] HCA-309',
-    name: 'HC 아트 우든체어',
-    deliveryNum: "158290398",
-    deliveryData: 
-      [
-        {
-            "location": "해외배송",
-            "deliveryStatus": "집하",
-            "date": "3/14(화)  06:33"
-        },
-        {
-            "location": "이탈리아 > 인천항",
-            "deliveryStatus": "배송중",
-            "date": "3/15(수)  01:14"
-        },
-        {
-            "location": "통관",
-            "deliveryStatus": "통관절차 진행중",
-            "date": "3/15(수)  08:21"
-        },
-        {
-            "location": "통관",
-            "deliveryStatus": "통관 처리완료",
-            "date": "3/15(수)  08:21"
-        },
-        {
-            "location": "국내배송",
-            "deliveryStatus": "배송중",
-            "date": "3/15(수)  14:02"
-        },
-        {
-            "location": "판교",
-            "deliveryStatus": "배송완료",
-            "date": "3/15(수)  17:11"
-        }
-    ]
-  },
-]
-
-MyDeliveryInfo.forEach((info) => {
+myDeliveryInfo.forEach((info) => {
   info.deliveryData.sort((a, b) => new Date(b.date).getTime() - new Date(a.date).getTime());
 });
 
@@ -106,9 +21,9 @@ const DeliveryDetail = () => {
       <Box sx={{ height: 'calc(100vh - 131px)', overflow: 'auto' }}>
         <Box sx={{ p: 2 }}>
           {
-            MyDeliveryInfo.map((el, index) => (
+            myDeliveryInfo.map((el, index) => (
               <Box key={index}>
-                <OrderCard>
+                <DeliveryInfoOrderCard>
                   <CardHeader title={el.date} />
                   <div className="card_content">
                     <p className="state">{el.state}</p>
@@ -155,15 +70,15 @@ const DeliveryDetail = () => {
                       CJ대한통운
                     </OutlineButton>
                   </CardActions>
-                </OrderCard>
+                </DeliveryInfoOrderCard>
               </Box>
             ))
           }
           
           <Box sx={{ mt: 3, mb: 5 }}>
-            {MyDeliveryInfo[0].deliveryData.map((item, index) => (
+            {myDeliveryInfo[0].deliveryData.map((item, index) => (
               <Box sx={{ display: "flex" }} key={index}>
-                {MyDeliveryInfo[0].deliveryData.length === 1 ? (
+                {myDeliveryInfo[0].deliveryData.length === 1 ? (
                   <>
                     <Box sx={{position:"relative"}}>
                       <IconDeliveryCompleted style={{position:"absolute", top:4,left:"-2px"}}/>
@@ -195,7 +110,7 @@ const DeliveryDetail = () => {
                     <Typography sx={{ fontSize: '14px', fontWeight: '700', color:"#000000", lineHeight:"20px" }}>{item.date}</Typography>
                   </Box>
                 </>
-                ) : index === MyDeliveryInfo[0].deliveryData.length - 1 ? (
+                ) : index === myDeliveryInfo[0].deliveryData.length - 1 ? (
                   <>
                     <Box sx={{position:"relative"}}>
                       <Box sx={{ml:"5.6px", height: "5px",borderLeft: "1px solid #D9D9D9"}}></Box>
