+++ src/pages/mypage/Review.tsx
@@ -2,61 +2,10 @@ import React, { useState, } from "react"
 import { Box, Tabs, Tab, } from "@mui/material"
 import NavigationBar from "../../components/NavigationBar"
 import CenterTitle from "../../components/title/CenterTitle"
-import { chair03 } from "../../assets/images/product"
 import OrderList from "./OrderList"
 import ReviewList from "./ReviewList"
-
-interface TabPanelProps {
-  children?: React.ReactNode
-  index: number
-  value: number
-}
-
-const myOrderList = [
-  {
-    date: '2023.02.23',
-    state: '배송완료',
-    img: chair03,
-    brand: 'Aerobiey',
-    optionCode: '[23s] AB-045',
-    name: '블랙 모던 소파',
-  },
-  {
-    date: '2022.03.13',
-    state: '배송완료',
-    img: chair03,
-    brand: 'Aerobiey',
-    optionCode: '[23s] AB-045',
-    name: '블랙 모던 소파',
-  },
-]
-
-function ReviewProps(index: number) {
-  return {
-    id: `Review-tab-${index}`,
-    'aria-controls': `Review-tabpanel-${index}`,
-  }
-}
-
-function ReviewTabPanel(props: TabPanelProps) {
-  const { children, value, index, ...other } = props
-
-  return (
-    <div
-      role="tabpanel"
-      hidden={value !== index}
-      id={`Review-tabpanel-${index}`}
-      aria-labelledby={`Review-tab-${index}`}
-      {...other}
-    >
-      {value === index && (
-        <Box>
-          {children}
-        </Box>
-      )}
-    </div>
-  )
-}
+import { TabPanel, TabProps } from "../../components/Tabs"
+import { reviewMyOrderData } from "../../data"
 
 const title = ['리뷰'];
 
@@ -83,22 +32,22 @@ const Review: React.FC = () => {
             onChange={handleReviewTab} 
             aria-label="main category"
           >
-            <Tab label="작성 가능한 리뷰" {...ReviewProps(0)} />
-            <Tab label="작성한 리뷰" {...ReviewProps(1)} />
+            <Tab label="작성 가능한 리뷰" {...TabProps(0)} />
+            <Tab label="작성한 리뷰" {...TabProps(1)} />
           </Tabs>
         </Box>
 
-        <ReviewTabPanel value={tabValue} index={0}>
+        <TabPanel value={tabValue} index={0}>
           <Box sx={{p:2}}>
-            <OrderList myOrderList={myOrderList} review={true} />
+            <OrderList myOrderList={reviewMyOrderData} review={true} />
           </Box>
-        </ReviewTabPanel>
+        </TabPanel>
 
-        <ReviewTabPanel value={tabValue} index={1}>
+        <TabPanel value={tabValue} index={1}>
           <Box sx={{p:2}}>
             <ReviewList />
           </Box>
-        </ReviewTabPanel>
+        </TabPanel>
       </Box>
       <NavigationBar />
     </>
