+++ src/pages/mypage/ReviewList.tsx
@@ -2,109 +2,12 @@ import React, { useState, } from "react"
 import { Box, Typography, Button, FormControl } from "@mui/material"
 import { deliveryProduct } from "../../assets/images/product"
 import { IconReviewStarBlackSmall, IconReviewStarGraySmall } from "../../assets/images/"
-import { styled } from "@mui/material/styles"
 import { useNavigate } from "react-router-dom"
-
-const RectCheckbox = styled(FormControl)(({ }) => ({
-  '& input': {
-    display: 'none',
-  },
-  '& label': {
-    padding: '6px 8px',
-    border: '1px solid #DADADA',
-    fontSize: '12px',
-    color: '#999999',
-  }
-}))
-
-const reviewList = [
-  {
-    date: '2023.02.23',
-    img: deliveryProduct,
-    brand: 'Aerobiey',
-    optionCode: '[23s] AB-045',
-    name: '블랙 모던 소파',
-    score: 4,
-    tagList: [
-      {
-        name: "생각보다 커요",
-      },
-      {
-        name: "색감 예뻐요",
-      },
-      {
-        name: "가벼워요",
-      },
-      {
-        name: "인스타감성",
-      },
-      {
-        name: "북유럽",
-      },
-      {
-        name: "분위기 있어요",
-      },
-    ],
-    contents: "배송은 한달 반정도 걸렸어요. 홈페이지에 안내된 것보다 조금더 늦어질 것같다는 안내는 업체에서 미리 주셨구요. 배송에 관한 문의전화를 했을때도 친절히 안내해주셨습니다."
-  },
-  {
-    date: '2023.02.23',
-    img: deliveryProduct,
-    brand: 'Aerobiey',
-    optionCode: '[23s] AB-045',
-    name: '블랙 모던 소파',
-    score: 3,
-    tagList: [
-      {
-        name: "생각보다 커요",
-      },
-      {
-        name: "색감 예뻐요",
-      },
-      {
-        name: "가벼워요",
-      },
-      {
-        name: "인스타감성",
-      },
-    ],
-    contents: "배송은 한달 반정도 걸렸어요. 홈페이지에 안내된 것보다 조금더 늦어질 것같다는 안내는 업체에서 미리 주셨구요. 배송에 관한 문의전화를 했을때도 친절히 안내해주셨습니다."
-  },
-  {
-    date: '2023.02.23',
-    img: deliveryProduct,
-    brand: 'Aerobiey',
-    optionCode: '[23s] AB-045',
-    name: '블랙 모던 소파',
-    score: 5,
-    tagList: [
-      {
-        name: "생각보다 커요",
-      },
-      {
-        name: "색감 예뻐요",
-      },
-      {
-        name: "가벼워요",
-      },
-      {
-        name: "인스타감성",
-      },
-      {
-        name: "북유럽",
-      },
-      {
-        name: "분위기 있어요",
-      },
-    ],
-    contents: "배송은 한달 반정도 걸렸어요."
-  },
-]
-
-
+import { RectCheckboxNonChecked } from "../../styles/checkbox.styles"
+import { reviewData } from "../../data"
 
 const ReviewList: React.FC = () => {
-  const [myReviewList, setMyReviewList] = useState(reviewList)
+  const [myReviewList, setMyReviewList] = useState(reviewData)
   
   const renderStarIcons = (score: number) => {
     const maxScore = 5; // 최대 점수
@@ -149,13 +52,13 @@ const ReviewList: React.FC = () => {
         <Typography sx={{color:"#333333", fontSize:14, lineHeight:"20px", letterSpacing:"-0.25px", mb:1, fontWeight:"700"}}>상품 태그</Typography>
         <Box sx={{mb:2}}>
           {review.tagList.map((el, index) => (
-            <RectCheckbox key={index} sx={{ mr: 1, mb: 1 }}>
+            <RectCheckboxNonChecked key={index} sx={{ mr: 1, mb: 1 }}>
               <input
                 type="checkbox"
                 id={`check_${index}`}
               />
               <label htmlFor={`check_${index}`}>{el.name}</label>
-            </RectCheckbox>
+            </RectCheckboxNonChecked>
           ))}
         </Box>
         <Typography sx={{color:"#333333", fontSize:14, lineHeight:"20px", letterSpacing:"-0.25px", mb:1, fontWeight:"700"}}>상품 후기</Typography>
