+++ src/pages/notification/index.tsx
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from "react";
 import CenterTitle from "../../components/title/CenterTitle";
 import NoData from "../../components/NoData";
 import ListItem from "./ListItem";
+import { notificationData } from "../../data";
 
 interface NotificationData {
   title: string;
@@ -16,30 +17,9 @@ const NotificationList: React.FC = () => {
   const linkText = "홈으로 가기";
   const linkTo = "/home";
 
-  const testData: NotificationData[] = [
-    {
-      title: "혜택",
-      name: "",
-      content: "앱 설치 후 회원가입 시 모든 브랜드 10% 할인쿠폰 증정",
-      date: "2023.09.01 00:00",
-    },
-    {
-      title: "답변",
-      name: "ALFDN",
-      content: "카멜프든 상품에 남겨주신 문의글에 답변이 달렸습니다.",
-      date: "2023.08.25 00:00",
-    },
-    {
-      title: "재입고알림",
-      name: "",
-      content: "[22 Series] OR-214 원레인 컬러드 체어가 재입고 되었습니다.",
-      date: "2023.09.01 21:40",
-    },
-  ];
-
   useEffect(() => {
     // 초기 데이터를 기존 데이터에 추가
-    setData([...data, ...testData]);
+    setData([...data, ...notificationData]);
   }, []); 
 
   return (
