+++ src/pages/onboarding/StyleSelection.tsx
@@ -6,7 +6,7 @@ import { useNavigate } from "react-router-dom"
 import TitleTwoLine2 from '../../components/title/TitleTwoLine2'
 import { GridBox } from "../../components/Box"
 import { StyleTypes } from "../shop/Filters/filters"
-import { StyleCard, StyleCardSelected } from "./StyleSelection.styles"
+import { StyleCard, StyleCardSelected } from "../../styles/card.style"
 import { IconLikeOn } from "../../assets/images"
 
 const StyleSelection = () => {
