+++ src/pages/order/OrderInfomation.tsx
@@ -1,11 +1,14 @@
 import React, { useState } from "react"
-import { Box, Avatar, List, ListItemAvatar, ListItemText, ListItemButton, Accordion, AccordionDetails, AccordionSummary, FormControl, Input } from "@mui/material"
+import { Box, Avatar, List, ListItemAvatar, ListItemText, ListItemButton, AccordionDetails, AccordionSummary, Input } from "@mui/material"
 import SelectBox from "../../components/SelectBox"
 import { styled } from "@mui/material/styles"
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore"
 import { PrimaryButton } from "../../styles/buttons.styles"
-import { aerobiey, jameslee, onerain, clods } from "../../assets/images/logo"
 import { IconDeliveryNormal, IconDeliveryQuick } from "../../assets/images"
+import { OrderAccordionMenu } from "../../styles/accordion.styles"
+import { DeliveryCheckBox } from "../../styles/checkbox.styles"
+import { OrderCardSelectWrap } from "../../styles/card.style"
+import { orderCouponData } from "../../data"
 
 interface OptionsProps {
   id?: string | undefined
@@ -23,135 +26,6 @@ const AvatarList = styled(List)(() => ({
     fontSize: '.875rem',
   },
 }))
-
-export const AccordionMenu = styled(Accordion)(() => ({
-  margin: '16px 16px 0',
-  boxShadow: 'none',
-  borderBottom: '1px solid #dadada',
-  '&.Mui-expanded': {
-    margin: '16px 16px 0',
-  },
-  '&:before': {
-    backgroundColor: 'transparent',
-  },
-  '& .MuiAccordionSummary-root': {
-    padding: 0,
-    '&.Mui-expanded': {
-      minHeight: 0,
-    },
-    '& > div': {
-      margin: '16px 0',
-      justifyContent: 'space-between',
-    },
-    '& > div.Mui-expanded': {
-      margin: '16px 0',
-    },
-    '& p': {
-      color: '#999',
-    },
-    '& .title, & .total_amount': {
-      color: '#333',
-      fontWeight: 'bold',
-    },
-  },
-  '& .MuiAccordionDetails-root': {
-    padding: '8px 0 16px',
-    borderTop: '1px solid #333',
-  }
-}))
-
-const DeliveryCheckBox = styled(FormControl)(() => ({
-  margin: '0 0 16px',
-  '& > input': {
-    display: 'none',
-  },
-  '& > label': {
-    display: 'flex',
-    alignItems: 'center',
-    padding: '16px',
-    height: '84px',
-    boxSizing: 'border-box',
-    border: '1px solid #dadada',
-  },
-  '& > label svg': {
-    color: '#999',
-  },
-  '& > label .price': {
-    fontSize: '1rem',
-  },
-  '& > label .additional': {
-    fontSize: '.875rem',
-    color: '#999',
-    letterSpacing: '-0.05rem',
-  },
-  '& > label .additional > span': {
-    color: '#FF6737',
-  },
-  '& input:checked + label': {
-    border: '1px solid #999',
-    backgroundColor: '#242223',
-    color: '#fafafa',
-  },
-  '& input:checked + label svg': {
-    color: '#fff',
-  }
-}))
-
-const CardSelectWrap = styled(FormControl)(() => ({
-  display: 'flex',
-  flexDirection: 'row',
-  height: '52px',
-
-  '& > .MuiBox-root, & .MuiButtonBase-root': {
-    display: 'flex',
-  },
-  '& > .MuiBox-root': {
-    flex: '1 1 auto',
-  },
-  '& .MuiInputBase-root': {
-    width: '100%',
-  },
-  '& .MuiInputBase-root > input': {
-    fontSize: '.875rem',
-  },
-  '& .MuiButtonBase-root': {
-    flex: '1 1 18%',
-    padding: 0,
-    borderRadius: 0,
-    fontSize: '.875rem',
-  },
-}))
-
-const couponItems = [
-  {
-    id: 'rerobiey',
-    brand: 'Rerobiey',
-    title: '브랜드 10% 쿠폰',
-    logo: aerobiey,
-    expiration: 20,
-  },
-  {
-    id: 'onerain',
-    brand: 'ONERAIN',
-    title: '브랜드 10% 쿠폰',
-    logo: onerain,
-    expiration: 15,
-  },
-  {
-    id: 'james_lee',
-    brand: 'James Lee',
-    title: '브랜드 10% 쿠폰',
-    logo: jameslee,
-    expiration: 11,
-  },
-  {
-    id: 'clods',
-    brand: 'CLODS',
-    title: '브랜드 10% 쿠폰',
-    logo: clods,
-    expiration: 3,
-  }
-]
 
 const OrderInfomation = () => {
   const [expanded, setExpanded] = useState<string | false>(false)
@@ -184,7 +58,7 @@ const OrderInfomation = () => {
 
   return (
     <>
-      <AccordionMenu
+      <OrderAccordionMenu
         expanded={expanded === 'panel1'} 
         onChange={handleChange('panel1')}
       >
@@ -202,7 +76,7 @@ const OrderInfomation = () => {
             setOpen={setOpen}
           >
             <AvatarList>
-              {couponItems.map((el) => (
+              {orderCouponData.map((el) => (
                 <ListItemButton
                   component="li"
                   key={el.id}
@@ -223,9 +97,9 @@ const OrderInfomation = () => {
             </AvatarList>
           </SelectBox>
         </AccordionDetails>
-      </AccordionMenu>
+      </OrderAccordionMenu>
 
-      <AccordionMenu
+      <OrderAccordionMenu
         expanded={expanded === 'panel2'} 
         onChange={handleChange('panel2')}
       >
@@ -275,9 +149,9 @@ const OrderInfomation = () => {
             </label>
           </DeliveryCheckBox>
         </AccordionDetails>
-      </AccordionMenu>
+      </OrderAccordionMenu>
 
-      <AccordionMenu
+      <OrderAccordionMenu
         expanded={expanded === 'panel3'} 
         onChange={handleChange('panel3')}
       >
@@ -288,7 +162,7 @@ const OrderInfomation = () => {
           <p>신용카드</p>
         </AccordionSummary>
         <AccordionDetails>
-          <CardSelectWrap
+          <OrderCardSelectWrap
             variant="standard"
             margin="normal"
           >
@@ -306,11 +180,11 @@ const OrderInfomation = () => {
             >
               선택
             </PrimaryButton>
-          </CardSelectWrap>
+          </OrderCardSelectWrap>
         </AccordionDetails>
-      </AccordionMenu>
+      </OrderAccordionMenu>
 
-      <AccordionMenu
+      <OrderAccordionMenu
         expanded={expanded === 'panel4'} 
         onChange={handleChange('panel4')}
       >
@@ -348,7 +222,7 @@ const OrderInfomation = () => {
             <p>1,435,050원</p>
           </Box>
         </AccordionDetails>
-      </AccordionMenu>
+      </OrderAccordionMenu>
     </>
   )
 }
