+++ src/pages/style/RankingList.tsx
@@ -1,151 +1,51 @@
-import React, { useState } from "react"
-import { Box, FormControl, Typography, Checkbox, Button } from "@mui/material"
+import React, { useState, useEffect } from "react"
+import { Box, Typography } from "@mui/material"
 import LeftTitleNoIcon from '../../components/title/LeftTitleNoIcon'
-import { rankingItem01, rankingItem02, rankingItem03, rankingItem04, rankingItem05, onerain01, onerain02 } from "../../assets/images/product"
-import { Link } from "react-router-dom";
 import ListItemType2 from "../../components/product/ListItemType2"
-import { styled } from "@mui/material/styles"
 import NavigationBar from "../../components/NavigationBar"
+import {rankingProductData, styleTypes} from "../../data/index"
+import {RectCheckbox} from "../../styles/checkbox.styles"
 
-
-export const rankingData = [
-  {
-    id: 1,
-    imgsrc: rankingItem01,
-  },
-  {
-    id: 2,
-    imgsrc: rankingItem02,
-  },
-  {
-    id: 3,
-    imgsrc: rankingItem03,
-  },
-  {
-    id: 4,
-    imgsrc: rankingItem04,
-  },
-  {
-    id: 5,
-    imgsrc: rankingItem05,
-  }, 
-]
-
-export const brandItemData = [
-  {
-    id: 0,
-    rank: 1,
-    imgsrc: rankingItem01,
-    shopName: 'ONERAIN',
-    modelName: '[23Series] OR-102',
-    productName: '원레인 우든 테이블',
-  },
-  {
-    id: 1,
-    rank: 2,
-    imgsrc: rankingItem02,
-    shopName: 'ONERAIN',
-    modelName: '[22Series] OR-210',
-    productName: '원레인 그랑프리 체어',
-  },
-  {
-    id: 2,
-    rank: 3,
-    imgsrc: rankingItem03,
-    shopName: 'ONERAIN',
-    modelName: '[23Series] OR-110',
-    productName: '원레인 우든 테이블',
-  },
-  {
-    id: 3,
-    rank: 4,
-    imgsrc: rankingItem04,
-    shopName: 'ONERAIN',
-    modelName: '[22Series] OR-214',
-    productName: '원레인 커러드 체어',
-  },
-  {
-    id: 4,
-    rank: 5,
-    imgsrc: rankingItem05,
-    shopName: 'ONERAIN',
-    modelName: '[23Series] OR-217',
-    productName: '원레인 럭커드 화이트 체어',
-  },
-]
-
-const itemList = [
-  {
-    id: 'sofa',
-    name: '소파',
-  },
-  {
-    id: 'bed',
-    name: '침대',
-  },
-  {
-    id: 'desk',
-    name: '책상',
-  },
-  {
-    id: 'table',
-    name: '테이블',
-  },
-  {
-    id: 'chair',
-    name: '의자',
-  },
-  {
-    id: 'lights',
-    name: '조명',
-  },
-  {
-    id: 'kids',
-    name: '키즈',
-  },
-  {
-    id: 'closet',
-    name: '수납장',
-  },
-  {
-    id: 'home_deco',
-    name: '홈데코',
-  }
-]
-
-const RectCheckbox = styled(FormControl)(({ }) => ({
-  '& input': {
-    display: 'none',
-  },
-  '& label': {
-    padding: '6px 16px',
-    border: '1px solid #DADADA',
-    fontSize: '0.875rem',
-    color: '#999999',
-  },
-  '& input:checked + label': {
-    border: '1px solid #242223',
-    backgroundColor: '#242223',
-    color: '#fff',
-  },
-}))
+const itemList = styleTypes.map(style => ({
+  id: (styleTypes.indexOf(style) + 1).toString(),
+  name: style.name,
+}));
 
 const RankingList = () => {
-
   const [filterItem, setFilterItem] = useState<string[]>([])
+  const [isAllChecked, setIsAllChecked] = useState(true);
+  const [filteredStyles, setFilteredStyles] = useState(rankingProductData);
 
-  const handleCheckedFilterItem = (e: React.ChangeEvent<HTMLInputElement>, id: string) => {
-    if (id === 'all') {
+  useEffect(() => {
+    setFilterItem(itemList.map(el => el.name));
+  }, []);
+
+  useEffect(() => {
+    const updatedFilteredStyles = rankingProductData.filter((style: any) => filterItem.includes(style.name));
+    setFilteredStyles(updatedFilteredStyles);
+  }, [filterItem]);
+
+  useEffect(() => {
+    if (filterItem.length === styleTypes.length) {
+      setIsAllChecked(true);
+    } else {
+      setIsAllChecked(false);
+    }
+  }, [filterItem]);
+
+
+  const handleCheckedFilterItem = (e: React.ChangeEvent<HTMLInputElement>, name: string) => {
+    if (name === 'all') {
       if (e.currentTarget.checked) {
-        setFilterItem(itemList.map(el => el.id))
+        setFilterItem(itemList.map(el => el.name))
       } else {
         setFilterItem([])
       }
     } else {
       if (e.currentTarget.checked) {
-        setFilterItem([...filterItem, id])
+        setFilterItem([...filterItem, name])
       } else {
-        setFilterItem(filterItem.filter(el => el !== id))
+        setFilterItem(filterItem.filter(el => el !== name))
       }
     }
   }
@@ -168,33 +68,34 @@ const RankingList = () => {
             type="checkbox"
             id={`check_all`}
             onChange={e => handleCheckedFilterItem(e, 'all')}
+            checked={isAllChecked}
           />
           <label htmlFor={`check_all`}>전체</label>
         </RectCheckbox>
         {itemList.map((el) => (
-          <RectCheckbox key={el.id} sx={{ ml: 1 }}>
+          <RectCheckbox key={el.name} sx={{ ml: 1 }}>
             <input
               type="checkbox"
-              id={`check_${el.id}`}
-              onChange={e => handleCheckedFilterItem(e, el.id)}
-              checked={filterItem.includes(el.id)}
+              id={el.name}
+              onChange={e => handleCheckedFilterItem(e, el.name)}
+              checked={filterItem.includes(el.name)}
             />
-            <label htmlFor={`check_${el.id}`}>{el.name}</label>
+            <label htmlFor={el.name}>{el.name}</label>
           </RectCheckbox>
         ))}
       </Box>
       <Box sx={{px: 2, minHeight: 'calc(100vh - 213.6px)'}}>
         {/* 가구리스트 */}
-        {brandItemData.map((product: any, index) =>
+        {filteredStyles.length != 0 ? (filteredStyles.map((product: any, index: any) =>
         <ListItemType2
           key={index}
-          rank={product.rank}
+          rank={index + 1}
           imgsrc={product.imgsrc}
           shopName={product.shopName}
           modelName={product.modelName}
           productName={product.productName}
         />
-      )}
+      )) : (<Typography sx={{ fontSize: 16, fontWeight: "500", letterSpacing: "-0.25px" }}>데이터가 없습니다</Typography>)}
       </Box>
       <NavigationBar/>
     </>
