+++ src/pages/style/RankingTabPanel.tsx
@@ -1,112 +1,12 @@
 import React, { useState } from "react"
-import { Box, FormControl, Typography, Checkbox, Button } from "@mui/material"
+import { Box, Typography } from "@mui/material"
 import HorizontalRanking  from "../../components/product/HorizontalRanking"
-import { rankingItem01, rankingItem02, rankingItem03, rankingItem04, rankingItem05, onerain01, onerain02 } from "../../assets/images/product"
-import { onerain, unknown } from "../../assets/images/brand"
 import { Link } from "react-router-dom";
 import ListItemType2 from "../../components/product/ListItemType2"
+import {rankingProductData} from "../../data"
 import { Swiper, SwiperSlide } from "swiper/react"
-
-export const rankingData = [
-  {
-    id: 1,
-    imgsrc: rankingItem01,
-  },
-  {
-    id: 2,
-    imgsrc: rankingItem02,
-  },
-  {
-    id: 3,
-    imgsrc: rankingItem03,
-  },
-  {
-    id: 4,
-    imgsrc: rankingItem04,
-  },
-  {
-    id: 5,
-    imgsrc: rankingItem05,
-  }, 
-]
-
-
-export const testData = [
-  {
-    id: 0,
-    imgsrc: onerain,
-    shopName: 'ONERAIN',
-    productList: [
-      {
-        id: 0,
-        imgsrc: onerain01,
-        shopName: 'ONERAIN',
-        modelName: '[22Series] OR-210',
-        productName: '원레인 그랑프리 체어',
-      },
-      {
-        id: 1,
-        imgsrc: onerain02,
-        shopName: 'ONERAIN',
-        modelName: '[23Series] OR-110',
-        productName: '원레인 우든 테이블',
-      },
-    ],
-  },
-  {
-    id: 1,
-    imgsrc: onerain,
-    shopName: 'ONERAIN',
-    productList: [
-      {
-        id: 0,
-        imgsrc: onerain01,
-        shopName: 'ONERAIN',
-        modelName: '[22Series] OR-210',
-        productName: '원레인 그랑프리 체어2',
-      },
-      {
-        id: 1,
-        imgsrc: onerain02,
-        shopName: 'ONERAIN',
-        modelName: '[23Series] OR-110',
-        productName: '원레인 우든 테이블2',
-      },
-    ],
-  },
-  {
-    id: 3,
-    imgsrc: onerain,
-    shopName: 'ONERAIN',
-    productList: [
-      {
-        id: 0,
-        imgsrc: onerain01,
-        shopName: 'ONERAIN',
-        modelName: '[22Series] OR-210',
-        productName: '원레인 그랑프리 체어3',
-      },
-      {
-        id: 1,
-        imgsrc: onerain02,
-        shopName: 'ONERAIN',
-        modelName: '[23Series] OR-110',
-        productName: '원레인 우든 테이블3',
-      },
-    ],
-  },
-]
-
-type ISwiperProps = { 
-  loop?: boolean
-  spaceBetween?: number 
-  scrollbar?: { 
-    draggable: boolean 
-    el: null 
-  }
-  slidesPerView?: "auto" | number | undefined
-  onSlideChange?: (swiper: any) => void
-}
+import { brandRakingSwiperData } from "../../data"
+import { ISwiperProps } from "../../components/ISwiperProps"
 
 const brandCollectionSettings:ISwiperProps = {
   loop: true,
@@ -120,12 +20,12 @@ const RankingTabPanel = () => {
     <>
       <Box sx={{mt:5}}>
         <Link to="/style/ranking"style={{ textDecoration: 'none', color: 'inherit' }}>
-          <HorizontalRanking data={rankingData} title="TOP5 가구랭킹" subTitle="MZ 세대를 위한 가구" discription= "심플한 디자인을 선호하는 MZ세대를 위한 가구"/>
+          <HorizontalRanking data={rankingProductData} title="TOP5 가구랭킹" subTitle="MZ 세대를 위한 가구" discription= "심플한 디자인을 선호하는 MZ세대를 위한 가구"/>
         </Link>
         <Box sx={{mt:"60px", ml:2, mb:2}}>
           <Typography sx={{ fontSize: 22, fontWeight: "700", color:"#333333", letterSpacing: "-1px" }}>실시간 인기 브랜드</Typography>
           <Swiper {...brandCollectionSettings}>
-          {testData.map((item, index) => (
+          {brandRakingSwiperData.map((item, index) => (
             <SwiperSlide key={index}>
               {({ isActive, isNext, isPrev }) => (
                 <Box >
